Enquiry client master chages completed

+++ iTSoft.CRM.UI/src/app/enquiry/enquiry-form.component.ts
@@ -22,26 +22,28 @@ import { Observable } from "rxjs/Observable";
 import { Subscription } from "rxjs/Subscription";
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 
+import { ConfirmDialog } from "../shared/dialog.component";
+import { MatDialog } from "@angular/material/dialog";
 import { CommandEventArgs, CommandModel, CommandType, ITMatTableComponent, TableColumnModel, TableDefaultSettings, ToolBarItems } from '../shared/table-layout/it-mat-table.component';
 
-import { RequestMaster } from '../_models';
+import { AddFollowupComponent } from '../process/add-followup/add-followup.component';
+import { AddServiceComponent } from "../process/add-service/add-service.component";
+
 import { RequestViewModel } from '../_models/requestviewmodel';
 import { RequestSelectListModel } from '../_models/requestselectlistmodel';
 import { RequestType } from '../_models/requesttype';
+import { ClientMaster } from "../masters/client/client.model";
+import { LeadStatus } from "../_models/leadStatus";
+import { RequestServiceDetails } from "../_models/requestservice";
 
 import { RequestService } from '../process/services/request.service';
 import { AlertService, StorageService } from '../_services';
 import { ListService } from '../process/services/list.service';
+import { UserProfilService } from "../_services/userProfile.Service";
+import { ContactPersonMaster } from "../_models/contactPerson";
 
-import { AddFollowupComponent } from '../process/add-followup/add-followup.component';
 
-import { NumberValidators } from "../shared/number.validator";
-import { GenericValidator } from "../shared/generic-validator";
-import { ConfirmDialog } from "../shared/dialog.component";
-import { MatDialog } from "@angular/material/dialog";
-import { AddServiceComponent } from "../process/add-service/add-service.component";
-import { RequestServiceDetails } from "../_models/requestservice";
-import { ClientMaster } from "../masters/client/client.model";
+
 
 
 @Component({
@@ -107,7 +109,7 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
     private requestService: RequestService,
     private listService: ListService,
     private alertService: AlertService,
-
+    private userProfileService: UserProfilService,
     private breakpointObserver: BreakpointObserver,
     private dialog: MatDialog,
   ) {
@@ -163,7 +165,7 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
     }
   }
 
-  getRequest(requestId: number , clientId : number): void {
+  getRequest(requestId: number, clientId: number): void {
     if (requestId > 0) {
       this.requestService
         .Load(requestId)
@@ -186,16 +188,13 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
       if (clientId > 0) {
         this.requestService.FindClient(clientId).subscribe((result) => {
           var clientDetails = <ClientMaster>result.Value.ResponseData;
-          if (clientDetails.CorporateName != null)
-            this.request.RequestMaster.CompanyName = clientDetails.CorporateName;
-          else if (clientDetails.FirstName != null)
-            this.request.RequestMaster.FirstName = clientDetails.LastName;
-          this.request.RequestMaster.MiddleName = clientDetails.MiddleName;
-          this.request.RequestMaster.LastName = clientDetails.LastName;
-          this.request.RequestMaster.DOB = clientDetails.DoB;
-
-          this.request.RequestMaster.Email = clientDetails.Email;
-          this.request.RequestMaster.PhoneNo1 = clientDetails.MobileNo;
+          if (clientDetails.CorporateName != null) {
+            this.request.OrganizationMaster.OrganizationName = clientDetails.CorporateName;
+            this.request.OrganizationMaster.OrganizationId = clientDetails.OrganizationId;
+          }
+          else if (clientDetails.FirstName != null) {
+            //this.request.ContactPersonMasters.push(clientDetails);
+          }
 
           this.request.RequestMaster.SourceId = 2;
 
@@ -209,6 +208,17 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
   onEnquiryRetrieved(request: RequestViewModel): void {
 
     this.request = request;
+
+    if(this.request.RequestMaster.ClientTypeId)
+    {
+      this.request.RequestMaster.ClientTypeId=this.request.RequestMaster.ClientTypeId.toString();
+    }
+    if (this.request.ContactPersonMasters && this.request.ContactPersonMasters.length > 0) {
+      this.request.ContactPersonMaster = this.request.ContactPersonMasters[0];
+    }
+    else {
+      this.request.ContactPersonMaster = new ContactPersonMaster();
+    }
 
     if (this.request.RequestMaster.RequestId == undefined || this.request.RequestMaster.RequestId === 0) {
       this.pageTitle = "Add Enquiry";
@@ -254,26 +264,41 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
     }
     else {
       let rowData: RequestServiceDetails = Object.assign({}, $event.rowData);
-      if ($event.command.commandType == CommandType.Edit) {
-        this.OpenServiceDialog(rowData);
-      }
-      else if ($event.command.commandType == CommandType.Delete) {
+      if (!(rowData.LeadStatusId == LeadStatus.ProposalAccepted || rowData.LeadStatusId == LeadStatus.Converted || rowData.LeadStatusId == LeadStatus.Dropped)) {
+        this.userProfileService.IsUserDepartment(rowData.DepartmentId).subscribe((result) => {
+          if (result == true) {
+            if ($event.command.commandType == CommandType.Edit) {
+              this.OpenServiceDialog(rowData);
+            }
+            else if ($event.command.commandType == CommandType.Delete) {
 
-        let dialogData = { title: "Confirm Action", message: "Are you sure ? Do you really want to Delete selected service ? " };
-        const dialogRef = this.dialog.open(ConfirmDialog, {
-          maxWidth: "400px",
-          data: dialogData
+              let dialogData = { title: "Confirm Action", message: "Are you sure ? Do you really want to Delete selected service ? " };
+              const dialogRef = this.dialog.open(ConfirmDialog, {
+                maxWidth: "400px",
+                data: dialogData
+              });
+
+              dialogRef.afterClosed().subscribe(dialogResult => {
+                let result = dialogResult;
+                if (result == "CONFIRMED") {
+                  this.requestService.RemoveService(rowData, this.request.RequestServiceDetails);
+                  this.serviceTable.RefreshDataSource();
+                }
+              }
+              );
+
+            }
+          }
+          else 
+          {
+            this.alertService.showErrorMessage("You are not authorize to modify service of this department");
+          }
         });
 
-        dialogRef.afterClosed().subscribe(dialogResult => {
-          let result = dialogResult;
-          if (result == "CONFIRMED") {
-                this.requestService.RemoveService(rowData,this.request.RequestServiceDetails);
-                this.serviceTable.RefreshDataSource();
-          }
-        }
-        );
 
+      }
+      else {
+        this.alertService.showWarningMessage("Action not allowed");
       }
 
     }
@@ -295,9 +320,8 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
         if (serviceDetails == null) {
           this.request.RequestServiceDetails.push(result.Data);
         }
-        else 
-        {
-            this.requestService.UpdateService(result.Data,this.request.RequestServiceDetails);
+        else {
+          this.requestService.UpdateService(result.Data, this.request.RequestServiceDetails);
         }
 
         this.request.RequestMaster.Amount = 0;
@@ -317,7 +341,7 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   onFollowUpSaved() {
-    this.getRequest(this.request.RequestMaster.RequestId,this.request.RequestMaster.ClientId);
+    this.getRequest(this.request.RequestMaster.RequestId, this.request.RequestMaster.ClientId);
   }
 
   LoadSelectListData() {
@@ -333,7 +357,7 @@ export class EnquiryFormComponent implements OnInit, AfterViewInit, OnDestroy {
             params => {
               let id = +params['id'];
               let clientId = +params['clientId'];
-              this.getRequest(id,clientId);
+              this.getRequest(id, clientId);
             }
           );
         }
